Fixing release build

+++ Engine/Source/ResourceSystem/Script.cpp
@@ -40,7 +40,7 @@ char* Engine::ResourceSystem::Script::ReadBytes(const std::filesystem::path& fil
 
     std::streampos end = stream.tellg();
     stream.seekg(0, std::ios::beg);
-    uint32_t size = end - stream.tellg();
+    uint32_t size = (uint32_t)end - (uint32_t)stream.tellg();
 
     if (size == 0)
     {
